refactor(badge-risk): compress badges into one component
