--- src/app/features/heat-map/modal-question-details/badge-risk-level/badge-risk-level.component.html
@@ -1,13 +1,0 @@
-@if (item) {
-  <div
-    class="risk-badge"
-    [style.backgroundColor]="getRiskColor(item.answerRisk)"
-    [style.color]="getTextRiskColor(item.answerRisk)"
-  >
-    {{ item.answerRisk }}
-  </div>
-} @else {
-  <app-empty-data-message description="No evaluation process found" />
-}
-
-<!-- (keyup.enter)="openModalDetails(item)" -->
