--- src/app/features/student/student-detail-option/badge-risk-level/badge-risk-level.component.html
@@ -1,13 +1,0 @@
-@if (item) {
-  <div
-    class="risk-badge"
-    [style.backgroundColor]="getRiskColor(item.riskSum)"
-    [style.color]="getTextRiskColor(item.riskSum)"
-  >
-    {{ item.riskSum }}
-  </div>
-} @else {
-  <app-empty-data-message description="No evaluation process found" />
-}
-
-<!-- (keyup.enter)="openModalDetails(item)" -->
