--- src/app/features/student/student-detail-option/badge-risk-level/badge-risk-level.component.scss
@@ -1,10 +1,0 @@
-.risk-badge {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 50%;
-  color: white;
-  width: 3em;
-  height: 3em;
-  margin: auto;
-}
