+++ src/app/shared/components/badge-person/badge-person.component.ts
