+++ src/app/shared/components/badge-risk-level/badge-risk-level.component.scss
