+++ src/app/shared/components/badge-risk-level/badge-risk-level.component.ts
@@ -1,12 +1,11 @@
 import { CommonModule } from '@angular/common';
 import { Component, Input } from '@angular/core';
 import { MatIconModule } from '@angular/material/icon';
-import { EmptyDataComponent } from '../../../../shared/components/empty-data/empty-data.component';
-import { StudentRiskResponse } from '../../../../core/models/cohort.model';
+import { EmptyDataComponent } from '../empty-data/empty-data.component';
 import {
   getRiskColor,
   getRiskTextColor,
-} from '../../../../core/constants/riskLevel';
+} from '../../../core/constants/riskLevel';
 
 @Component({
   selector: 'app-badge-risk-level',
@@ -15,7 +14,7 @@ import {
   imports: [MatIconModule, CommonModule, EmptyDataComponent],
 })
 export class BadgeRiskComponent {
-  @Input() item: StudentRiskResponse | undefined;
+  @Input() riskLevel: number | undefined;
 
   getRiskColor(riskLevel: number): string {
     return getRiskColor(riskLevel);
